cleaning up my comments on pg templates & js
before this, i had notes that talked through my decisions and explained
why i did certain things.  i might want to revert to this if i can for
building new templates.

with this branch, i’m cleaning those up for the devs that need to
rework this.  sorry, it’s not bootstrapped, but it’s simple enough to
troubleshoot in my opinion.